Changes: 	• Made a mtg constants file, not sure if I'll keep it 	• Segmented out MTG files into a directory.
Plans Moving Forward:

Reacting Plans
1:
	Turn current queries
		○ Get Card
		○ GetRuling
		○ GetSets

		○ GetQuery (when it's finished)

	Into more up to standard services that work appropriately.

2:
	Fix that god forsaken Dice rolling implementation.  It's horrible, I know.

3:
	Segment our "flags" to be on one file and we import and manipulate that way.

4:
	Rewrite our key to be better.

Current Actions:
	• Fully implement our query functionality (even if you miss edgecases right now)
	• To the drawing board for the HIGH LEVEL STRUCTURE
