+++ go.mod
@@ -3,7 +3,6 @@ module goland-discord-bot
 go 1.18
 
 require (
-	github.com/MagicTheGathering/mtg-sdk-go v0.0.0-20190109105601-3aaea97721aa // indirect
 	github.com/bwmarrin/discordgo v0.25.0 // indirect
 	github.com/gorilla/websocket v1.4.2 // indirect
 	golang.org/x/crypto v0.0.0-20210421170649-83a5a9bb288b // indirect
